+++ database/migrations/2018_04_10_054535_create_orders_table.php
@@ -17,7 +17,7 @@ class CreateOrdersTable extends Migration
             $table->increments('id');
             $table->integer('reagent_id')->unsigned();
             $table->integer('person_id')->unsigned();
-            $table->integer('order_quantity');
+            $table->float('order_quantity', 8, 2);
             $table->integer('state')->default(0);
             $table->date('handed_date');
             $table->timestamps();
